+++ modules/aws-kernel/test/src/smithy4s/aws/CryptoTest.scala
@@ -28,7 +28,7 @@ object CryptoTest extends FunSuite {
     val expected =
       "f7bc83f430538424b13298e6aa6fb143ef4d59a14946175997479dbc2d1a3cd8"
     val result = toHexString(hmacSha256(data, binaryFromString(key)))
-    assert(expected == result)
+    expect(expected == result)
   }
 
 }