+++ modules/tests/src/smithy4s/tests/PizzaSpec.scala
@@ -334,7 +334,7 @@ abstract class PizzaSpec
           log
         )
         .map(_._1)
-        .map(assert.eql(_, 400))
+        .map(expect.eql(_, 400))
   }
 
   routerTest("query param failing refinement results in a BadRequest") {
@@ -348,7 +348,7 @@ abstract class PizzaSpec
           log
         )
         .map(_._1)
-        .map(assert.eql(_, 400))
+        .map(expect.eql(_, 400))
   }
 
   routerTest("body failing refinement results in a BadRequest") {
@@ -361,7 +361,7 @@ abstract class PizzaSpec
           log
         )
         .map(_._1)
-        .map(assert.eql(_, 400))
+        .map(expect.eql(_, 400))
   }
 
   routerTest("Optional payload set to empty") { (client, uri, log) =>